@@ -155,7 +155,7 @@ extern "C" {
     /** Reverts an FLSharedKeys by "forgetting" any keys added since it had the count `oldCount`. */
     FLEECE_PUBLIC void FLSharedKeys_RevertToCount(FLSharedKeys, unsigned oldCount) FLAPI;
 
-    /** Disable caching of the SharedKeys.. */
+    /** Disable caching of the SharedKeys. */
     FLEECE_PUBLIC void FLSharedKeys_DisableCaching(FLSharedKeys) FLAPI;
 
     /** Increments the reference count of an FLSharedKeys. */
