@@ -23,22 +23,18 @@
 #endif
 #include "betterassert.hh"
 
+#pragma clang diagnostic ignored "-Wmissing-noreturn"
+
 // Catch's REQUIRE is too slow for perf testing
 #undef REQUIRE
 #define REQUIRE(TEST)   while (!(TEST)) {abort();}
 #undef CHECK
 #define CHECK(TEST)     while (!(TEST)) {abort();}
-#ifdef __APPLE__
-#define FLEECE_UNUSED __unused
-#elif !defined(_MSC_VER)
-#define FLEECE_UNUSED __attribute__((unused))
-#else
-#define FLEECE_UNUSED
-#endif
 
 
 using namespace fleece;
 using namespace fleece::impl;
+using namespace fleece_test;
 
 
 #if !FL_EMBEDDED
@@ -116,7 +112,7 @@ TEST_CASE("Perf LoadFleece", "[.Perf]") {
         Benchmark bench;
         for (int i = 0; i < kIterations; i++) {
             bench.start();
-            FLEECE_UNUSED auto root = Value::fromData(doc)->asArray();
+            [[maybe_unused]] auto root = Value::fromData(doc)->asArray();
             REQUIRE(root != nullptr);
             bench.stop();
         }
@@ -130,7 +126,7 @@ TEST_CASE("Perf LoadFleece", "[.Perf]") {
         for (int i = 0; i < kIterations; i++) {
             bench.start();
             for (int j = 0; j < kIterationsPerSample; j++) {
-                FLEECE_UNUSED auto root = Value::fromTrustedData(doc)->asArray();
+                [[maybe_unused]] auto root = Value::fromTrustedData(doc)->asArray();
                 REQUIRE(root != nullptr);
             }
             bench.stop();
