@@ -11,8 +11,10 @@
 //
 
 #include "fleece/FLBase.h"
+#include "fleece/InstanceCounted.hh"
 #include "FleeceTests.hh"
 #include "FleeceImpl.hh"
+#include "Backtrace.hh"
 #include "ConcurrentMap.hh"
 #include "Bitmap.hh"
 #include "TempArray.hh"
@@ -22,9 +24,31 @@
 #include <future>
 
 using namespace std;
+using namespace fleece;
 
 
-// TESTS:
+// Tests for the LIFETIMEBOUND attribute on slices. This code is commented out because it should fail to compile.
+#if 0
+static slice bad(const char* str) {
+    alloc_slice strslice(str);
+    return strslice;            // Clang should flag this as an error thanks to LIFETIMEBOUND
+}
+
+static slice bad2(const char* str) {
+    string s(str);
+    return s;                   // Clang should flag this as an error thanks to LIFETIMEBOUND
+}
+
+
+TEST_CASE("Slice safety") {
+    // Both of these calls will result in use-after-free UB:
+    slice x = bad("long string will be heap allocated, right?");
+    CHECK(string(x) == "foo");
+    x = bad2("long string will be heap allocated, right?");
+    CHECK(string(x) == "foo");
+}
+#endif
+
 
 #ifndef _MSC_VER
 
@@ -343,4 +367,36 @@ TEST_CASE("Timestamp Conversions", "[Timestamps]") {
     FLStringResult str = FLTimestamp_ToString(ts, asUTC);
     FLTimestamp ts2 = FLTimestamp_FromString((FLString)str);
     CHECK(ts == ts2);
+    FLSliceResult_Release(str);
+}
+
+
+static Backtrace makeBacktrace() {
+    return Backtrace();
+}
+
+
+TEST_CASE("Backtrace") {
+    Backtrace bt = makeBacktrace();
+    string str = bt.toString();
+    cout << str << endl;
+    CHECK(std::count(str.begin(), str.end(), '\n') >= 4);
+}
+
+
+class ICTest : public RefCounted, public InstanceCountedIn<ICTest> {
+public:
+    ICTest() noexcept = default;
+};
+
+
+TEST_CASE("InstanceCounted") {
+    auto n = InstanceCounted::liveInstanceCount();
+    {
+        auto i1 = make_retained<ICTest>();
+        auto i2 = make_retained<ICTest>();
+        InstanceCounted::dumpInstances();
+        CHECK(InstanceCounted::liveInstanceCount() == n + 2);
+    }
+    CHECK(InstanceCounted::liveInstanceCount() == n);
 }
